Refactor generic code into `cratedb_toolkit` namespace
