+++ cratedb_toolkit/exception.py
@@ -1,0 +1,2 @@
+class TableNotFound(Exception):
+    pass
