+++ cratedb_toolkit/util/__init__.py
@@ -1,0 +1,3 @@
+from cratedb_toolkit.util.cli import boot_with_dburi  # noqa: F401
+from cratedb_toolkit.util.data import jd  # noqa: F401
+from cratedb_toolkit.util.database import DatabaseAdapter  # noqa: F401
