+++ cratedb_toolkit/util/cli.py
@@ -8,7 +8,7 @@ import typing as t
 
 import click
 
-from cratedb_retention.util.common import setup_logging
+from cratedb_toolkit.util.common import setup_logging
 
 logger = logging.getLogger(__name__)
