+++ cratedb_toolkit/util/common.py
