+++ cratedb_toolkit/util/data.py
