+++ examples/edit.py
@@ -30,12 +30,11 @@ Synopsis
 import logging
 import os
 
-from cratedb_retention.model import DatabaseAddress, JobSettings, RetentionPolicy, RetentionStrategy
+from cratedb_retention.model import JobSettings, RetentionPolicy, RetentionStrategy
 from cratedb_retention.setup.schema import setup_schema
 from cratedb_retention.store import RetentionPolicyStore
-from cratedb_retention.util.cli import boot_with_dburi
-from cratedb_retention.util.data import jd
-from cratedb_retention.util.database import DatabaseAdapter
+from cratedb_toolkit.model import DatabaseAddress
+from cratedb_toolkit.util import DatabaseAdapter, boot_with_dburi, jd
 
 logger = logging.getLogger(__name__)
