+++ tests/conftest.py
@@ -2,14 +2,15 @@
 # Distributed under the terms of the AGPLv3 license, see LICENSE.
 import pytest
 
-from cratedb_retention.model import DatabaseAddress, JobSettings, RetentionPolicy, RetentionStrategy
+from cratedb_retention.model import JobSettings, RetentionPolicy, RetentionStrategy
 from cratedb_retention.setup.schema import setup_schema
 from cratedb_retention.store import RetentionPolicyStore
-from cratedb_retention.util.common import setup_logging
-from cratedb_retention.util.database import DatabaseAdapter, run_sql
-from tests.testcontainers.azurite import ExtendedAzuriteContainer
-from tests.testcontainers.cratedb import CrateDBContainer
-from tests.testcontainers.minio import ExtendedMinioContainer
+from cratedb_toolkit.model import DatabaseAddress
+from cratedb_toolkit.testing.testcontainers.azurite import ExtendedAzuriteContainer
+from cratedb_toolkit.testing.testcontainers.cratedb import CrateDBContainer
+from cratedb_toolkit.testing.testcontainers.minio import ExtendedMinioContainer
+from cratedb_toolkit.util.common import setup_logging
+from cratedb_toolkit.util.database import DatabaseAdapter, run_sql
 
 # Use different schemas both for storing the retention policy table, and
 # the test data, so that they do not accidentally touch the default `doc`
